Renamed into GraviSuitOverlay to fit more
